Бизнес логика создана и успешно функционирует

+++ app/src/main/java/com/authapp/api/paymentlist/Payment.kt
@@ -1,6 +1,7 @@
 package com.authapp.api.paymentlist
 
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 import java.math.BigInteger
 
 @JsonClass(generateAdapter = true)
@@ -8,5 +9,5 @@ data class Payment(
     val id: Int,
     val title: String,
     val amount: Any?,
-    val created: BigInteger?
+    val created: Long?
 )
